defer all playing to the playground
