+++ demo/package.json
@@ -11,6 +11,7 @@
   "//": "the codemirror fork is due to https://github.com/JedWatson/react-codemirror/issues/106#issuecomment-302674960",
   "dependencies": {
     "@skidding/react-codemirror": "1.0.1",
+    "dedent-js": "^1.0.1",
     "js-yaml": "^3.8.4",
     "lodash": "^4.17.4",
     "neutrino-middleware-postcss": "^2.0.2",
